Finish select, refactor
